split implementation for customer, change package name
