+++ internal/customer/models/customer.go
@@ -1,4 +1,4 @@
-package customer
+package models
 
 import "Dzaakk/simple-commerce/package/template"
